requirements.txt
PySide6==6.6.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
bcrypt==4.0.1
python-dotenv==1.0.0
reportlab==4.0.4
openpyxl==3.1.2
pytest==7.4.3
black==23.11.0